refactor(constant): 상수 파일 분리
많은 상수가 필요해져서 더 쉬운 유지보수를 위해 상수 성격에 맞추어 파일들을 분리
