+++ src/constant/inputMessage.js
@@ -1,0 +1,5 @@
+export const INPUT_MESSAGE = Object.freeze({
+  PRICE: '구입금액을 입력해 주세요.\n',
+  WINNING_NUMBERS: '당첨 번호를 입력해 주세요.\n',
+  BONUS_NUMBER: '보너스 번호를 입력해 주세요.\n',
+});
