+++ src/constant/system.js
@@ -1,0 +1,21 @@
+export const MATCH_PRICE = Object.freeze({
+  THREE: 5000,
+  FOUR: 50000,
+  FIVE: 1500000,
+  FIVE_BONUS: 30000000,
+  SIX: 2000000000,
+});
+
+export const NUMBER_RANGE = Object.freeze({
+  MIN: 1,
+  MAX: 45,
+});
+
+export const PRICE_RANGE = Object.freeze({
+  MIN: 1000,
+  MAX: 1000000,
+});
+
+export const LOTTO_NUMBER_LENGTH = 6;
+
+export const SEPARATOR = ',';
